requirements.txt
django==3.0.7
django-debug-toolbar==2.2
dj-database-url==0.5.0
Pillow==7.1.2
environs==9.3.0
django-phonenumber-field==5.0.0
phonenumbers==8.12.19
djangorestframework==3.12.2
geopy==2.1.0
equests=2.25.1